Added ecgAnalysis.m function
[ P, Q, R, S, T ] = ecgAnalysis( ecg, Fs, det ) extract the main ECG
features by ?rst detecting the R peaks and their positions in the
signal, then we can extract the other peaks (P, Q, S, T) respectively ,
relative to the R-peaks. R-peaks are detected by first taking the detail
coefficients of the original signal via 2nd level wavelet decomposition.
This is the squared and the positions of the maxima in a sliding window
of size almost equal to one pulse of the ecg signal is recorded. The
other peaks are located by finding the maximas and minimas in the
locality of the R peaks in the approximation coefficeint signal.